requirements.txt
nltk==3.8.1
gensim==4.3.3
requests
PatternLite
fastapi[standard]==0.112.1
sqlalchemy==2.0.32
EbookLib==0.18
beautifulsoup4==4.12.3
openai==1.78.0
pydantic
pyjwt
bcrypt==4.0.1
passlib[bcrypt]